% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_drug_annot}
\alias{add_drug_annot}
\title{Annotates tidy plate data to drug metadata.}
\usage{
add_drug_annot(tidy_data_df, drug_key_df)
}
\arguments{
\item{tidy_data_df}{Tidied and gathered long-format data tibble produced by
\code{gather_plates()}}

\item{drug_key_df}{A tibble key of corresponding (\code{catalog_no}, \code{cas_no},
\code{drug_name}) to (\code{plate}, \code{plate_row}, \code{plate_col}). All values are
mandatory except for \code{catalog_no} as the rest of the information is needed
for joining or used as downstream metadata.}
}
\value{
tibble
}
\description{
Tidy plate data produced by \code{gather_plates()} is joined to drug_key metadata
based on the combination of 'plate', \code{plate_row}, and \code{plate_col}.
}
\examples{
\dontrun{
# Get inputs
result_paths <- Sys.glob("data/raw_id_corrected/*.CSV")
drug_key_df <- readr::read_csv("data/config/drug_well_key_complete.csv")

# Process data
batch_data <- gather_plates(result_paths)
annot_data <- add_drug_annot(batch_data$tidy_data, drug_key_df)
}
}
