% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{agg_data}
\alias{agg_data}
\title{Aggregates plate reader .CSVs and annotates wells with drug data.}
\usage{
agg_data(paths, drug_key_df)
}
\arguments{
\item{paths}{string vector of paths to csv files.}

\item{drug_key_df}{dataframe of well information corresponding to drug info.}
}
\value{
A list with $data and $metadata tibbles.
}
\description{
loops through csv paths, tidies experimental data and metadata,
adds drug information to data, and aggregates all data in a list of
experimental data and metadata
}
\examples{
paths <- Sys.glob("data/raw_id_corrected/*.CSV")
drug_key_df <- readr::read_csv("data/config/drug_well_key_complete.csv")
test <- agg_data(paths, drug_key_df)
}
