% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpacaqc-package.R
\docType{package}
\name{alpacaqc-package}
\alias{alpacaqc}
\alias{alpacaqc-package}
\title{alpacaqc: ALPACA QC}
\description{
A package for tidying plate reader data and adding drug annotation information from ALPACA (Australian Library of Paediatric Anti-Cancer Agents) drug libraries.
}
\author{
\strong{Maintainer}: Vikesh Ajith \email{vikesh.ajith@hudson.org.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
