% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gather_plates}
\alias{gather_plates}
\title{Gathers plate reader csv data into tidied long-format.}
\usage{
gather_plates(result_paths)
}
\arguments{
\item{result_paths}{string vector of paths to csv files output by plate
readers in a standard human-readable format.}
}
\value{
A list with $tidy_data and $tidy_metadata tibbles.
}
\description{
loops through csv paths, extracts experimental data, and extracts metadata
using regex.
}
\examples{
result_paths <- Sys.glob("data/raw_id_corrected/*.CSV")
test <- gather_plates(result_paths)
}
