% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{import_drug_key}
\alias{import_drug_key}
\title{Imports and validates a drug_key csv file}
\usage{
import_drug_key(drug_key_path)
}
\arguments{
\item{drug_key_path}{A path to a csv correlating well locations (\verb{[plate, plate_row, plate_col]}) to drug information (\verb{[catalog_no, cas_no, drug_name]}).

Well information should be unique across \verb{[plate, plate_row, plate_col]}
and follow the specifications of a 384 well plate (i.e plate_row and
plate_col should fall within ranges A-P and 1-24, respectively)

All fields should be complete (excepting the \code{cas_no} for \code{PURO/PBS})}
}
\value{
A tibble() that can then be supplied to add_drug_annot()
}
\description{
Imports and validates a drug_key csv file
}
\examples{
\dontrun{
drug_key_df <- import_drug_key("data/config/drug_well_key_complete.csv")
}
}
