% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plate_heatmap}
\alias{plate_heatmap}
\title{Visualises plate data as a heatmap}
\usage{
plate_heatmap(tidy_data, value = "fluor")
}
\arguments{
\item{tidy_data}{A tidied tibble produced by \code{gather_plates()$tidy_data} or
\code{add_drug_annot()}. Ideally filtered to a few specific plates of interest,
otherwise the output plot will be very large due to faceting.}

\item{value}{A string corresponding to the heatmap measure of interest,
default \code{"fluor"}}
}
\value{
\code{ggplot()} object
}
\description{
Visualises plate data as a heatmap faceted by \code{treatment}, \code{rep} \code{conc_nm}
and \code{plate}
}
\examples{
\dontrun{
annot_data \%>\%
    filter(conc_nm == 10) \%>\%
    plot_plate()
}
}
